requirements.txt
azure-functions
python-dotenv
requests